refactor: split log messages into individual plugins
