+++ packages/analytics-js-plugins/src/bugsnag/index.ts
@@ -5,7 +5,7 @@ import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { IExternalSrcLoader } from '@rudderstack/analytics-js-common/services/ExternalSrcLoader/types';
 import { ExtensionPlugin } from '@rudderstack/analytics-js-common/types/PluginEngine';
 import { BugsnagLib } from '../types/plugins';
-import { BUGSNAG_API_KEY_VALIDATION_ERROR } from '../utilities/logMessages';
+import { BUGSNAG_API_KEY_VALIDATION_ERROR } from './logMessages';
 import { API_KEY } from './constants';
 import { initBugsnagClient, loadBugsnagSDK, isApiKeyValid, getAppStateForMetadata } from './utils';
