+++ packages/analytics-js-plugins/src/errorReporting/index.ts
@@ -8,7 +8,7 @@ import {
 import { IExternalSrcLoader } from '@rudderstack/analytics-js-common/services/ExternalSrcLoader/types';
 import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
-import { INVALID_SOURCE_CONFIG_ERROR } from '../utilities/logMessages';
+import { INVALID_SOURCE_CONFIG_ERROR } from './logMessages';
 
 const pluginName = 'ErrorReporting';
