+++ packages/analytics-js-plugins/src/errorReporting/logMessages.ts
@@ -1,0 +1,3 @@
+const INVALID_SOURCE_CONFIG_ERROR = `Invalid source configuration or source id.`;
+
+export { INVALID_SOURCE_CONFIG_ERROR };
