+++ packages/analytics-js-plugins/src/ketchConsentManager/index.ts
@@ -7,7 +7,7 @@ import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { isUndefined } from '@rudderstack/analytics-js-common/utilities/checks';
 import { DestinationConfig } from '@rudderstack/analytics-js-common/types/Destination';
 import { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandler';
-import { DESTINATION_CONSENT_STATUS_ERROR } from '../utilities/logMessages';
+import { DESTINATION_CONSENT_STATUS_ERROR } from './logMessages';
 import { KETCH_CONSENT_MANAGER_PLUGIN } from './constants';
 import { KetchConsentData } from './types';
 import { getKetchConsentData, updateConsentStateFromData } from './utils';
