+++ packages/analytics-js-plugins/src/oneTrustConsentManager/index.ts
@@ -7,7 +7,7 @@ import { ExtensionPlugin } from '@rudderstack/analytics-js-common/types/PluginEn
 import { IStoreManager } from '@rudderstack/analytics-js-common/types/Store';
 import { OneTrustCookieCategory } from '@rudderstack/analytics-js-common/types/Consent';
 import { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandler';
-import { DESTINATION_CONSENT_STATUS_ERROR, ONETRUST_ACCESS_ERROR } from '../utilities/logMessages';
+import { DESTINATION_CONSENT_STATUS_ERROR, ONETRUST_ACCESS_ERROR } from './logMessages';
 import { ONETRUST_CONSENT_MANAGER_PLUGIN } from './constants';
 import { OneTrustGroup } from './types';
