+++ packages/analytics-js-plugins/src/shared-chunks/index.ts
@@ -1,3 +1,2 @@
 export * as storages from '@rudderstack/analytics-js-common/constants/storages';
 export * as encryptionUtils from '../storageEncryption/encryptionUtils';
-export * as logMessages from '../utilities/logMessages';
