+++ packages/analytics-js-plugins/src/utilities/queue.ts
@@ -4,11 +4,18 @@ import { stringifyWithoutCircular } from '@rudderstack/analytics-js-common/utili
 import { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
 import { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import { Nullable } from '@rudderstack/analytics-js-common/types/Nullable';
+import { LOG_CONTEXT_SEPARATOR } from '@rudderstack/analytics-js-common/constants/logMessages';
 import { EVENT_PAYLOAD_SIZE_BYTES_LIMIT } from './constants';
-import {
-  EVENT_PAYLOAD_SIZE_CHECK_FAIL_WARNING,
-  EVENT_PAYLOAD_SIZE_VALIDATION_WARNING,
-} from './logMessages';
+
+const EVENT_PAYLOAD_SIZE_CHECK_FAIL_WARNING = (
+  context: string,
+  payloadSize: number,
+  sizeLimit: number,
+): string =>
+  `${context}${LOG_CONTEXT_SEPARATOR}The size of the event payload (${payloadSize} bytes) exceeds the maximum limit of ${sizeLimit} bytes. Events with large payloads may be dropped in the future. Please review your instrumentation to ensure that event payloads are within the size limit.`;
+
+const EVENT_PAYLOAD_SIZE_VALIDATION_WARNING = (context: string): string =>
+  `${context}${LOG_CONTEXT_SEPARATOR}Failed to validate event payload size. Please make sure that the event payload is within the size limit and is a valid JSON object.`;
 
 const QUEUE_UTILITIES = 'QueueUtilities';
